% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Methods_Related.R
\name{ClusterMethods}
\alias{ClusterMethods}
\title{Clusters outlier detection methods according to the correlation matrix for min-max and median-iqr normalization methods, using the R package \code{cluster}.}
\usage{
ClusterMethods(kk, rocpr = 1, vis = FALSE)
}
\arguments{
\item{kk}{The number of clusters}

\item{rocpr}{If \code{rocpr = 1}, then area under ROC cure is used, if \code{rocpr = 2} then area under PR curve is used.}

\item{vis}{If \code{TRUE} then the resulting graphs are plotted.}
}
\value{
A \code{pam.object} described in the R package \code{cluster}.
}
\description{
Clusters outlier detection methods according to the correlation matrix for min-max and median-iqr normalization methods, using the R package \code{cluster}.
}
\examples{
cls <- ClusterMethods(8)
which(cls$clustering==2)
which(cls$clustering==4)
which(cls$clustering==5)
which(cls$clustering==6)

}
\seealso{
\code{\link[cluster]{pam.object}}.
}
