% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Functions.R
\name{ComputeMetaFeaturesAll}
\alias{ComputeMetaFeaturesAll}
\title{Computes features from a dataset, but using 2 normalization methods: MIN_MAX and MEDIAN_IQR.}
\usage{
ComputeMetaFeaturesAll(dat.o)
}
\arguments{
\item{dat.o}{This is the data with the last column named as \code{outlier} having values \code{yes} or \code{no}. All other columns are numeric. Non-numerical data need to be converted to numerical before computing features.}
}
\value{
The set of features.
}
\description{
Computes features from a dataset, but using 2 normalization methods: MIN_MAX and MEDIAN_IQR.
}
\examples{
data(Arrhythmia_withoutdupl_05_v05)
dat <- Arrhythmia_withoutdupl_05_v05
feat <- ComputeMetaFeaturesAll(dat)

}
