% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Methods_Related.R
\name{CorrPrRoc}
\alias{CorrPrRoc}
\title{Computes correlation between AUCROC performance values and AUCPR performance values}
\usage{
CorrPrRoc(m = 1)
}
\arguments{
\item{m}{If \code{1}, then \code{pearson}, if \code{2}, then \code{kendall} or if \code{3}, then \code{spearman} correlation is performed. Default value is \code{1}.}
}
\value{
The correlation values for each performance column.
}
\description{
Computes correlation between AUCROC performance values and AUCPR performance values
}
\examples{
\dontrun{
corobj <- CorrPrRoc()
}

}
