% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Performance.R
\name{CrossValidateModels}
\alias{CrossValidateModels}
\title{Cross validates models to reproduce results in  journal and conference papers.}
\usage{
CrossValidateModels(d, p, s = 1, n = 5)
}
\arguments{
\item{d}{If \code{d=1} then we take the Min_Max performance values, if \code{d=2} then performance values from all normalization methods are considered. Input values for  \code{d} are only \code{1,2}.}

\item{p}{If \code{p=1} then we take binary values based on absolute performance, i.e. if performance \code{ > 0.8}, if \code{p=2} the relative binary performance values are used. Input values for  \code{p} are only \code{1,2}.}

\item{s}{If \code{s=1} then we train the models on a preferred subset of features. If \code{s=2} the models are trained on all features, which takes considerably longer.  Default value \code{s=1}. Input values for  \code{s} are only \code{1,2}.}

\item{n}{The number of folds in cross validation.}
}
\value{
A list containing the following results from the cross validated models:
\describe{
 \item{def_acc}{The default accuracy we get if we predict the method is not good for all instances. This is the percentage of the majority class.}
 \item{results}{The \code{n}-fold cross valdation results. }
 \item{mean_acc}{The mean \code{n}-fold cross valdation results.}
}
}
\description{
Cross validates models to reproduce results in  journal and conference papers.
}
\examples{
\dontrun{
out <- CrossValidateModels(1,1,1,5)
out$mean_acc
}

}
