% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Prediction.R
\name{CrossValidateSensitivityToNorm}
\alias{CrossValidateSensitivityToNorm}
\title{Cross validates models to predict sensitivity to normalization.}
\usage{
CrossValidateSensitivityToNorm(rocpr = 1, xi = 0.05, n = 10)
}
\arguments{
\item{rocpr}{If \code{rocpr=1} then area under ROC curve is used as the performance measure.  If \code{rocpr=2} area under PR curve is used.}

\item{xi}{The xi-sensitivity to normalization parameter xi. Defaults to \code{0.05}}

\item{n}{The number of folds in cross validation.}
}
\value{
A list containing the following results from the cross validated models:
\describe{
 \item{def_acc}{The default accuracy we get if we predict the method is not good for all instances. This is the percentage of the majority class.}
 \item{results}{The \code{n}-fold cross valdation results. }
 \item{mean_acc}{The mean \code{n}-fold cross valdation results.}
}
}
\description{
Cross validates models to predict sensitivity to normalization.
}
\examples{
\dontrun{
out <- CrossValidateSensitivityToNorm(1,0.05,10)
out$mean_acc
}

}
