% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Performance.R
\name{CrossValidateSVM}
\alias{CrossValidateSVM}
\title{Cross validates SVM with instance space coordinates.}
\usage{
CrossValidateSVM(d = 1, n = 5)
}
\arguments{
\item{d}{If \code{d=1} then we take the Min_Max performance values, if \code{d=2} then performance values from all normalization methods are considered. Input values for  \code{d} are only \code{1,2}.}

\item{n}{The number of folds in cross validation.}
}
\value{
A list with the following components:
\describe{
 \item{def_acc}{The default accuracy we get if we predict the method is not good for all instances. This is the percentage of the majority class.}
 \item{results}{The \code{n}-fold cross valdation results. }
 \item{mean_acc}{The mean \code{n}-fold cross valdation results.}
 \item{d}{Which instance space is cross validated. If \code{d=1}, then it is MIN_MAX normalization instance space, if \code{d=2} then all normalization methods are used. }
 \item{coordinates}{The instance space coordinates.}
}
}
\description{
Cross validates SVM with instance space coordinates.
}
\examples{
\dontrun{ out <- CrossValidateSVM(1,5) }

}
