% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Performance.R
\name{InstSpace}
\alias{InstSpace}
\title{Trains an SVM on instance space coordinates and draws the instance space.}
\usage{
InstSpace(d = 1, vis = FALSE)
}
\arguments{
\item{d}{If \code{d=1} then we take the Min_Max performance values, if \code{d=2} then performance values from all normalization methods are considered. Input values for  \code{d} are only \code{1,2}.}

\item{vis}{If \code{TRUE} then the instance space is plotted.}
}
\value{
A list with the following components:
\describe{
 \item{preds10}{The predictions as \code{1} or \code{0} for each instance and each algorithm.}
 \item{predprobs}{The prediction probabilities for each each instance and each algorithm.}
 \item{algorithms}{The algorithm best suited for each instance. For some instances, no algorithms work well.}
 \item{d}{Which instance space is cross validated. If \code{d=1}, then it is MIN_MAX normalization instance space, if \code{d=2} then all normalization methods are used. }
 \item{coordinates}{The instance space coordinates.}
}
}
\description{
Trains an SVM on instance space coordinates and draws the instance space.
}
\examples{
\dontrun{ svmout <- InstSpace(d=1)}

}
