% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Methods_Related.R
\name{IsMinMaxBetter}
\alias{IsMinMaxBetter}
\title{Performs \code{binom.test} to see if Min-Max is actually better than Median-IQR method.}
\usage{
IsMinMaxBetter(rocpr = 1, prop = 0.5)
}
\arguments{
\item{rocpr}{If \code{rocpr=1} then area under ROC curve is used as the performance measure.  If \code{rocpr=2} area under PR curve is used.}

\item{prop}{Claimed proportion that Min-Max is better. Used for the null hypothesis.}
}
\value{
A list containing the following:
\describe{
 \item{percentages}{The proportion of datasets that gave better performance for Min-Max.}
 \item{pvalues}{The p-value of the  \code{binom.test}.}
 \item{confintervals}{The confidence intervals of the  \code{binom.test}.}
}
}
\description{
Performs \code{binom.test} to see if Min-Max is actually better than Median-IQR method.
}
\examples{
out <- IsMinMaxBetter(1)
out$confintervals
}
