% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Performance.R
\name{PlotNewInstance}
\alias{PlotNewInstance}
\title{Plots a new instance in the instance space.}
\usage{
PlotNewInstance(svm_out, feat)
}
\arguments{
\item{svm_out}{The output of the trained svm using function \code{DrawInstSpace}}

\item{feat}{The features of the new instance/dataset. This can be computed using \code{ComputeMetaFeaturesAll} or \code{ComputeMetaFeaturesMM}.}
}
\value{
new_coords Coodinates of the new instance in the instance space
}
\description{
Plots a new instance in the instance space.
}
\examples{
data(Arrhythmia_withoutdupl_05_v05)
dat <- Arrhythmia_withoutdupl_05_v05
feat <- ComputeMetaFeaturesMM(dat)
svmout <- DrawInstSpace(d=1)
PlotNewInstance(svm_out, feat)

}
