% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Performance.R
\name{PredictPerformance}
\alias{PredictPerformance}
\title{Predicts suitable outlier methods from meta-features of the dataset using trained random forest models.}
\usage{
PredictPerformance(ftrs, models)
}
\arguments{
\item{ftrs}{Meta-features of the dataset.}

\item{models}{The trained random forest models. These are obtained from the function \code{TrainModels}.}
}
\value{
The prediction probabilities for each of the outlier method. For example, a probability of 0.78 for the first outlier method means the probability that the first method is good for this dataset is 0.78.
}
\description{
Predicts suitable outlier methods from meta-features of the dataset using trained random forest models.
}
\examples{
\dontrun{
data(Arrhythmia_withoutdupl_05_v05)
dat <- Arrhythmia_withoutdupl_05_v05
feat <- ComputeMetaFeaturesMM(dat)
fit <- TrainModels(1,1,1)
out <- PredictPerformance(feat, fit)
}

}
