% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Sensitivity.R
\name{SensitivityNormDatasets}
\alias{SensitivityNormDatasets}
\title{Computes sensitive to normalization for dataset sources}
\usage{
SensitivityNormDatasets(rocpr = 1, xi = 0.05)
}
\arguments{
\item{rocpr}{If \code{rocpr=1} then area under ROC curve is used as the performance measure.  If \code{rocpr=2} area under PR curve is used.}

\item{xi}{The xi-sensitivity to normalization parameter xi. Defaults to \code{0.05}}
}
\value{
A list containing the following:
\describe{
 \item{KruskalWallis}{The output of the KruskalWallis from \code{stats::kruskal.test}.}
 \item{max_val}{The dataset source, which is most sensitive to normalization.}
 \item{min_val}{The dataset source, which is least sensitive to normalization. }
 \item{xi_sensitive_num}{For how many outlier methods each dataset is sensitive to normalization.}
}
}
\description{
Computes sensitive to normalization for dataset sources
}
\examples{
out <- SensitivityNormDatasets(1, 0.05)
out

}
