% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Methods_Related.R
\name{SensitivityToNorm}
\alias{SensitivityToNorm}
\title{Computes percentages of datasets which are sensitive to normalization.}
\usage{
SensitivityToNorm(rocpr, xi)
}
\arguments{
\item{rocpr}{If \code{rocpr=1} then area under ROC curve is used as the performance measure.  If \code{rocpr=2} area under PR curve is used.}

\item{xi}{Sensitivity to normalization parameter.     For a given dataset, we say that an outlier detection method is \code{xi}-sensitive to normalization if the difference between the maximum performance and the minimum performance across all normalization schemes for that outlier detection method is greater than \code{xi}.}
}
\value{
A list containing the following:
\describe{
 \item{percentages}{The proportion of datasets that gave better performance for Min-Max.}
 \item{pvalues}{The p-value of the  \code{binom.test}.}
 \item{confintervals}{The confidence intervals of the  \code{binom.test}.}
 \item{methods}{The outlier detection methods.}
}
}
\description{
Computes percentages of datasets which are sensitive to normalization.
}
\examples{
out <- SensitivityToNorm(1,0.05)
out$confintervals

}
