% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Sensitivity.R
\name{SensitivityToNormMixedMod}
\alias{SensitivityToNormMixedMod}
\title{Fits a mixed effects model to account for normalization methods and outlier methods accounting for the dataset variants}
\usage{
SensitivityToNormMixedMod(rocpr = 1, ds = FALSE, pp = FALSE)
}
\arguments{
\item{rocpr}{If \code{rocpr=1} then area under ROC curve is used as the performance measure.  If \code{rocpr=2} area under PR curve is used.}
}
\value{
A list containing the following:
\describe{
 \item{fit1}{The first model without interactions between normalization and outlier method.}
 \item{fit2}{The second model with interactions between normalization and outlier method.}
 \item{aov}{The output of an ANOVA test. }
}
}
\description{
Fits a mixed effects model to account for normalization methods and outlier methods accounting for the dataset variants
}
\examples{
out <- SensitivityToNormMixedMod(1)
out

}
