% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Methods_Related.R
\name{SensitivityToNorm}
\alias{SensitivityToNorm}
\title{Computes sensitive to normalization statistics.}
\usage{
SensitivityToNorm(rocpr = 1)
}
\arguments{
\item{rocpr}{If \code{rocpr=1} then area under ROC curve is used as the performance measure.  If \code{rocpr=2} area under PR curve is used.}
}
\value{
A list containing the following:
\describe{
 \item{friedman}{The output of the Friedman test.}
 \item{nemenyi}{The output of the Nemenyi test \code{tsutils::nemenyi}.}
}
}
\description{
Computes sensitive to normalization statistics.
}
\examples{
out <- SensitivityToNorm(1)
out

}
